Exit after every get and the last getnext entry
It turns out that snmpd leaves the pass_persist script running, which makes
sense, but leaves the burden of updating the OID values to the script itself.
Otherwise the script will just return the initially cached values for all
eternity. So, brutally exit after each "get" and when then end of the OID
subtree is reached with "getnext" requests. A slightly cleaner way might be to
refresh the cached values after every "get" and "getnext" so that the script
would not have to exit at all.

Signed-off-by: Samuli Seppänen <samuli.seppanen@puppeteers.net>